move properties to properties dir
